refactor(web): modularize autopilot task maintenance
Split the autopilot task editor into smaller schedule and run-history modules so the UI is easier to evolve. Keep the desktop autopilot schema aligned with Prisma and document the scheduler's best-effort semantics.
